# fix: faster toString for integrity