@wraithgar 2023-04-03
`npm run lintfix` should shore up the linting errors (it usually fixes everything but line length warnings)