@wraithgar 2023-04-03
Reading the readme again we do not make any guarantees about the order that the algorithms will be in `toString`.  The order doesn't matter to things like `parse` either so I think this can remain a patch release.